Opinion securized

+++ src/Manager/OpinionManager.php
@@ -5,14 +5,17 @@ namespace App\Manager;
 use App\Entity\Opinion;
 use App\Entity\User;
 use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Component\Security\Core\Security;
 
 class OpinionManager
 {
     private EntityManagerInterface $entityManager;
+    private Security $security;
 
-    public function __construct(EntityManagerInterface $entityManager)
+    public function __construct(EntityManagerInterface $entityManager, Security $security)
     {
         $this->entityManager = $entityManager;
+        $this->security = $security;
     }
 
     public function findOrCreate(User $author, string $tmdbId): Opinion
@@ -37,7 +40,9 @@ class OpinionManager
     {
         $opinion = $this->findOrCreate($author, $tmdbId);
 
-        $opinion->setIsViewed(true);
+        if ($this->security->isGranted('edit', $opinion)) {
+            $opinion->setIsViewed(true);
+        }
 
         return $this->save($opinion);
     }
@@ -46,21 +51,27 @@ class OpinionManager
     {
         $opinion = $this->findOrCreate($author, $tmdbId);
 
-        $opinion->setIsViewed(false);
+        if ($this->security->isGranted('edit', $opinion)) {
+            $opinion->setIsViewed(false);
+        }
 
         return $this->save($opinion);
     }
 
     public function delete(Opinion $opinion): void
     {
-        $this->entityManager->remove($opinion);
-        $this->entityManager->flush();
+        if ($this->security->isGranted('delete', $opinion)) {
+            $this->entityManager->remove($opinion);
+            $this->entityManager->flush();
+        }
     }
 
     public function save(Opinion $opinion): Opinion
     {
-        $this->entityManager->persist($opinion);
-        $this->entityManager->flush();
+        if ($this->security->isGranted('edit', $opinion)) {
+            $this->entityManager->persist($opinion);
+            $this->entityManager->flush();
+        }
 
         return $opinion;
     }
